+++ changelog.md
@@ -5,6 +5,8 @@ Uiua is not yet stable.
 # 0.12.0 - 2024-??-??
 This version is not yet released. If you are reading this on the website, then these changes are live here.
 ### Language
+- **Breaking Change** - [`&runs`](https://uiua.org/docs/&runs) now returns 3 handles rather than 1
+  - This gives more precise control over stdin, stdout, and stderr
 - Add *tree extension*, which allows some functions that take indices, offsets, or masks to recursively on tree of those values
   - This is useful for working with boxed arrays
   - Currently implemented for [`keep ▽`](https://uiua.org/docs/keep), [`select ⊏`](https://uiua.org/docs/select), and [`pick ⊡`](https://uiua.org/docs/pick)